[REF] Refactor of package name
